add documentations for user interaction rearrange package for user interaction (screens -> interface adapter + frameworks)
