+++ src/main/java/abr/user_interact_abr/manage_friend_request_abr/deleting_attempt_abr/DeleteFriendOrDenyFriendRequest.java
@@ -18,6 +18,10 @@ public class DeleteFriendOrDenyFriendRequest implements FriendManagerInputBounda
         this.friendManagerPresenter = friendManagerPresenter;
     }
 
+    /**
+     * @param requestModel the input data containing user's ID (username), friend's ID and user's friendList
+     * @return a FriendManagerResponseModel that contains a msg to display on screen and an updated user's friendList
+     */
     @Override
     public FriendManagerResponseModel reactTo(FriendManagerRequestModel requestModel) {
 
@@ -31,7 +35,7 @@ public class DeleteFriendOrDenyFriendRequest implements FriendManagerInputBounda
 
         userDsGateway.save(requestModel.getUserID(), requestModel.getFriendID(), tempUserFriendList, tempFriendFriendList);
 
-        FriendManagerResponseModel responseModel = new FriendManagerResponseModel("You are no longer friends with " + requestModel.getFriendID(), tempUserFriendList);
+        FriendManagerResponseModel responseModel = new FriendManagerResponseModel("Deleted/Denied successfully", tempUserFriendList);
         return friendManagerPresenter.showMsgAndUpdatedFriendList(responseModel);
     }
 }
