+++ src/main/java/ds/user_interact_ds/FriendPlaylistFileDsGateway.java
@@ -1,4 +1,4 @@
-package ds.user_interact_DS;
+package ds.user_interact_ds;
 
 import abr.user_interact_abr.friend_playlist_interact_abr.FriendPlaylistDsGateway;
 import entities.playlist_entities.Playlist;
@@ -6,6 +6,9 @@ import entities.playlist_entities.Playlist;
 import java.util.ArrayList;
 
 public class FriendPlaylistFileDsGateway implements FriendPlaylistDsGateway {
+    /**
+     * TBA since playlist DB is not finalized
+     */
     @Override
     public ArrayList<Playlist> getUserPlaylist(String userName) {
         return null;
