+++ src/main/java/framework/user_interact_screen/friend_manager_screen/FriendListView.java
@@ -1,10 +1,17 @@
-package interface_adaptors.user_interact_screen.friend_manager_screen;
+package framework.user_interact_screen.friend_manager_screen;
+
+import interface_adaptors.user_interact_ia.DeleteFriendOrDenyFriendRequestController;
+import interface_adaptors.user_interact_ia.SendFriendRequestController;
+import interface_adaptors.user_interact_ia.ShowFriendListController;
 
 import javax.swing.*;
 import java.awt.*;
 import java.awt.event.ActionEvent;
 import java.awt.event.ActionListener;
 
+/**
+ * A screen that shows the friendList of current user
+ */
 public class FriendListView extends JPanel{
     ShowFriendListController showFriendListController;
     SendFriendRequestController acceptFriendRequestController;
