+++ src/main/java/interface_adaptors/user_interact_ia/CheckFriendPlaylistController.java
@@ -1,11 +1,14 @@
-package interface_adaptors.user_interact_screen.friend_playlist_interact_screen;
+package interface_adaptors.user_interact_ia;
 
 import abr.user_interact_abr.friend_playlist_interact_abr.CheckFriendPlaylistInputBoundary;
-import interface_adaptors.user_interact_screen.friend_manager_screen.TempFriendListObservable;
 
 import java.util.ArrayList;
 
 
+/**
+ * all user interaction controllers serve the same purpose: get input data from screen, pass input data to use cases,
+ * get output data from user cases, pass output data to screen.
+ */
 public class CheckFriendPlaylistController{
 
     final CheckFriendPlaylistInputBoundary inputBoundary;
@@ -14,7 +17,7 @@ public class CheckFriendPlaylistController{
         this.inputBoundary = inputBoundary;
     }
 
-    ArrayList<String> returnFilteredPlaylist(String friendName){
+    public ArrayList<String> returnFilteredPlaylist(String friendName){
         return inputBoundary.returnViewablePlaylist(TempFriendListObservable.currentUser, friendName);
     }
 }
