+++ src/test/java/user_interact_abr_test/SendFriendRequestTest.java
@@ -2,7 +2,7 @@ package user_interact_abr_test;
 
 import abr.user_interact_abr.manage_friend_request_abr.*;
 import abr.user_interact_abr.manage_friend_request_abr.sending_or_accepting_attempt_abr.SendFriendRequest;
-import ds.user_interact_DS.FriendManagerInMemoryDsGateway;
+import ds.user_interact_ds.FriendManagerInMemoryDsGateway;
 import org.junit.Test;
 
 import java.util.HashMap;
